+++ api/src/main/java/org/mapstruct/tools/gem/GemDefinitions.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.annotations.test;
+package org.mapstruct.tools.gem;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -13,8 +13,12 @@ import java.lang.annotation.Target;
 /**
  * @author Sjaak Derksen
  */
-@Retention(RetentionPolicy.SOURCE)
-@Target(ElementType.TYPE )
-public @interface Builder {
-    // deliberately empty
+@Retention(RetentionPolicy.CLASS)
+@Target({ ElementType.PACKAGE,ElementType.TYPE })
+public @interface GemDefinitions {
+
+    /**
+     * @return The gem definitions
+     */
+    GemDefinition[] value();
 }
