+++ api/src/main/java/org/mapstruct/tools/gem/GemValue.java
@@ -1,4 +1,4 @@
-package org.mapstruct.annotations;
+package org.mapstruct.tools.gem;
 
 import javax.lang.model.element.AnnotationValue;
