+++ processor/src/main/java/org/mapstruct/tools/gem/processor/GemValueInfo.java
@@ -1,4 +1,4 @@
-package org.mapstruct.annotations.processor;
+package org.mapstruct.tools.gem.processor;
 
 import javax.lang.model.type.TypeMirror;
