+++ processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.mapstruct.annotations.processor.GemProcessor
+org.mapstruct.tools.gem.processor.GemProcessor
