+++ test/src/main/java/org/mapstruct/tools/gem/test/Builder.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.annotations.test;
+package org.mapstruct.tools.gem.test;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -13,14 +13,8 @@ import java.lang.annotation.Target;
 /**
  * @author Sjaak Derksen
  */
-@Retention(RetentionPolicy.CLASS)
-@Target( ElementType.TYPE )
-public @interface SomeAnnotations {
-
-    /**
-     * The configuration of the bean attributes.
-     *
-     * @return The configuration of the bean attributes.
-     */
-    SomeAnnotation[] value();
+@Retention(RetentionPolicy.SOURCE)
+@Target(ElementType.TYPE )
+public @interface Builder {
+    // deliberately empty
 }
